requirements.txt
fastapi
uvicorn
pytest
httpx
torch
numpy
scipy
pandas
pyarrow
fastparquet
requests
python-dotenv
ratelimit
scikit-learn